Split GtkHelper/GtkHelper.py widget classes into widgets/
