+++ GtkHelper/widgets/BetterPreferencesGroup.py
@@ -1,0 +1,65 @@
+"""
+Author: Core447
+Year: 2023
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+any later version.
+
+This programm comes with ABSOLUTELY NO WARRANTY!
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <https://www.gnu.org/licenses/>.
+"""
+import gi
+
+gi.require_version("Gtk", "4.0")
+gi.require_version("Adw", "1")
+from gi.repository import Adw
+
+
+class BetterPreferencesGroup(Adw.PreferencesGroup):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def clear(self):
+        list_box = self.get_list_box()
+        list_box.remove_all()
+
+    def set_sort_func(self, *args, **kwargs):
+        list_box = self.get_list_box()
+        list_box.set_sort_func(*args, **kwargs)
+
+    def set_filter_func(self, *args, **kwargs):
+        list_box = self.get_list_box()
+        list_box.set_filter_func(*args, **kwargs)
+
+    def invalidate_filter(self):
+        list_box = self.get_list_box()
+        list_box.invalidate_filter()
+
+    def invalidate_sort(self):
+        list_box = self.get_list_box()
+        list_box.invalidate_sort()
+
+    def get_rows(self):
+        list_box = self.get_list_box()
+        if list_box is None:
+            return
+
+        rows = []
+        child = list_box.get_first_child()
+        while child is not None:
+            rows.append(child)
+            child = child.get_next_sibling()
+
+        return rows
+
+    def get_list_box(self):
+        first_box = self.get_first_child()
+        second_box = first_box.get_first_child()
+        third_box = second_box.get_next_sibling()
+        list_box = third_box.get_first_child()
+
+        return list_box
