+++ GtkHelper/widgets/EntryRowWithoutTitle.py
@@ -1,0 +1,34 @@
+"""
+Author: Core447
+Year: 2023
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+any later version.
+
+This programm comes with ABSOLUTELY NO WARRANTY!
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <https://www.gnu.org/licenses/>.
+"""
+import gi
+
+gi.require_version("Gtk", "4.0")
+gi.require_version("Adw", "1")
+from gi.repository import Adw, Gtk
+
+
+class EntryRowWithoutTitle(Adw.EntryRow):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        # Make title invisible
+        child:Gtk.Box = self.get_child() # Box
+        prefix_box:Gtk.Box = child.get_first_child()
+        gizmo = prefix_box.get_next_sibling()
+        empty_title = gizmo.get_first_child()
+        title = empty_title.get_next_sibling()
+
+        empty_title.set_visible(False)
+        title.set_visible(False)
