+++ GtkHelper/widgets/RevertButton.py
@@ -1,0 +1,25 @@
+"""
+Author: Core447
+Year: 2023
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+any later version.
+
+This programm comes with ABSOLUTELY NO WARRANTY!
+
+You should have received a copy of the GNU General Public License
+along with this program. If not, see <https://www.gnu.org/licenses/>.
+"""
+import gi
+
+gi.require_version("Gtk", "4.0")
+gi.require_version("Adw", "1")
+from gi.repository import Gtk
+
+
+class RevertButton(Gtk.Button):
+    def __init__(self, **kwargs):
+        super().__init__(icon_name="edit-undo-symbolic", **kwargs)
+        self.set_tooltip_text("Revert to action defaults")
